requirements.txt
django==2.2
mysqlclient
Pillow
django-simple-captcha
redis